requirements.txt
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.3
cryptography==36.0.1
dnspython==2.2.0
Flask==2.0.2
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
idna==3.3
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
MongoAlchemy==0.19
pycparser==2.21
PyJWT==2.3.0
pymodm==0.4.3
pymongo==3.12.3
requests==2.27.1
SQLAlchemy==1.4.31
urllib3==1.26.8
Werkzeug==2.0.3
gunicorn